api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy>=1.24
scipy>=1.10
sentence-transformers>=2.7